Session request: Future / roadmap for git-repo
Per [1, 2] there are assumptions about this. There was a presentation
by Conley Owens a couple of years ago, and Stefan Beller has been doing
a lot of work to improve submodule support, but is there a clear roadmap
for when (or if) repo will be phased out and replaced by submodules?

[1] https://groups.google.com/d/msg/repo-discuss/eUk8APuQovo/GAM5eO0nDQAJ
[2] https://code.google.com/p/git-repo/issues/detail?id=223

Change-Id: I8fdb602e1c756fa53a8d84c5aee7f549a8fb3418